+++ air-script/src/tests/buses/buses_varlen_boundary_last_plonky3.rs
@@ -1,7 +1,8 @@
-use p3_air::{Air, BaseAir, BaseAirWithPublicValues, AirBuilder, ExtensionBuilder};
+use p3_field::{ExtensionField, Field, PrimeCharacteristicRing};
 use p3_matrix::Matrix;
-use p3_field::{Field, PrimeCharacteristicRing};
-use crate::test_utils::plonky3_traits::{AirScriptAir, AirScriptBuilder};
+use p3_matrix::dense::RowMajorMatrixView;
+use p3_matrix::stack::VerticalPair;
+use p3_miden_air::{MidenAir, MidenAirBuilder, RowMajorMatrix};
 
 pub const MAIN_WIDTH: usize = 1;
 pub const AUX_WIDTH: usize = 2;
@@ -12,41 +13,75 @@ pub const MAX_BETA_CHALLENGE_POWER: usize = 2;
 
 pub struct BusesAir;
 
-impl<F> BaseAir<F> for BusesAir {
+impl<F, EF> MidenAir<F, EF> for BusesAir
+where F: Field,
+      EF: ExtensionField<F>,
+{
     fn width(&self) -> usize {
         MAIN_WIDTH
     }
-}
 
-impl<F> BaseAirWithPublicValues<F> for BusesAir {
-    fn num_public_values(&self) -> usize {
-        NUM_PUBLIC_VALUES
+    fn num_randomness(&self) -> usize {
+        1 + MAX_BETA_CHALLENGE_POWER
     }
-}
 
-impl<F: Field, AB: AirScriptBuilder<F = F>> AirScriptAir<F, AB> for BusesAir {
     fn aux_width(&self) -> usize {
         AUX_WIDTH
     }
 
-    fn max_beta_challenge_power(&self) -> usize {
-        MAX_BETA_CHALLENGE_POWER
+    fn build_aux_trace(&self, _main: &RowMajorMatrix<F>, _challenges: &[EF]) -> Option<RowMajorMatrix<EF>> {
+        // Note: consider using Some(build_aux_trace_with_miden_vm::<F, EF>(_main, _challenges, module)) if you want to build the aux trace using Miden VM aux trace builders.
+
+        let num_rows = _main.height();
+        let trace_length = num_rows * AUX_WIDTH;
+        let mut long_trace = EF::zero_vec(trace_length);
+        let mut trace = RowMajorMatrix::new(long_trace, AUX_WIDTH);
+        let (prefix, rows, suffix) = unsafe { trace.values.align_to_mut::<[EF; AUX_WIDTH]>() };
+        assert!(prefix.is_empty(), "Alignment should match");
+        assert!(suffix.is_empty(), "Alignment should match");
+        assert_eq!(rows.len(), num_rows);
+        // Initialize first row
+        let initial_values = Self::buses_initial_values::<F, EF>();
+        for j in 0..AUX_WIDTH {
+            rows[0][j] = initial_values[j];
+        }
+        // Fill subsequent rows using direct access to the rows array
+        for i in 0..num_rows-1 {
+            let i_next = (i + 1) % num_rows;
+            let main_local = _main.row_slice(i).unwrap(); // i < height so unwrap should never fail.
+            let main_next = _main.row_slice(i_next).unwrap(); // i_next < height so unwrap should never fail.
+            let main = VerticalPair::new(
+                RowMajorMatrixView::new_row(&*main_local),
+                RowMajorMatrixView::new_row(&*main_next),
+            );
+            let periodic_values: [_; NUM_PERIODIC_VALUES] = <BusesAir as MidenAir<F, EF>>::periodic_table(self).iter().map(|col| col[i % col.len()]).collect::<Vec<_>>().try_into().expect("Wrong number of periodic values");
+            let prev_row = &rows[i];
+            let next_row = Self::buses_transitions::<F, EF>(
+                &main,
+                _challenges,
+                &periodic_values,
+                prev_row,
+            );
+            for j in 0..AUX_WIDTH {
+                rows[i+1][j] = next_row[j];
+            }
+        }
+        Some(trace)
     }
 
-    fn periodic_table(&self) -> Vec<Vec<F>> {
-        vec![]
-    }
-
-    fn eval(&self, builder: &mut AB) {
+    fn eval<AB>(&self, builder: &mut AB)
+    where AB: MidenAirBuilder<F = F, EF = EF>,
+    {
         let public_values: [_; NUM_PUBLIC_VALUES] = builder.public_values().try_into().expect("Wrong number of public values");
         let periodic_values: [_; NUM_PERIODIC_VALUES] = builder.periodic_evals().try_into().expect("Wrong number of periodic values");
+        let preprocessed = builder.preprocessed();
         let main = builder.main();
         let (main_current, main_next) = (
             main.row_slice(0).unwrap(),
             main.row_slice(1).unwrap(),
         );
-        let alpha = builder.alpha();
-        let beta_challenges: [_; MAX_BETA_CHALLENGE_POWER] = builder.beta_powers().try_into().expect("Wrong number of beta challenges");
+        let (&alpha, beta_challenges) = builder.permutation_randomness().split_first().expect("Wrong number of randomness");
+        let beta_challenges: [_; MAX_BETA_CHALLENGE_POWER] = beta_challenges.try_into().expect("Wrong number of randomness");
         let aux_bus_boundary_values: [_; AUX_WIDTH] = builder.aux_bus_boundary_values().try_into().expect("Wrong number of aux bus boundary values");
         let aux = builder.permutation();
         let (aux_current, aux_next) = (
@@ -70,8 +105,31 @@ impl<F: Field, AB: AirScriptBuilder<F = F>> AirScriptAir<F, AB> for BusesAir {
     }
 }
 
-impl<AB: AirScriptBuilder> Air<AB> for BusesAir {
-    fn eval(&self, builder: &mut AB) {
-        <Self as AirScriptAir<AB::F, AB>>::eval(self, builder);
+impl BusesAir {
+    fn buses_initial_values<F, EF>() -> Vec<EF>
+    where F: Field,
+          EF: ExtensionField<F>,
+    {
+        vec![
+            EF::ONE,
+            EF::ZERO,
+        ]
+    }
+
+    fn buses_transitions<F, EF>(main: &VerticalPair<RowMajorMatrixView<F>, RowMajorMatrixView<F>>, challenges: &[EF], periodic_evals: &[F], aux_current: &[EF]) -> Vec<EF>
+    where F: Field,
+          EF: ExtensionField<F>,
+    {
+        let (main_current, main_next) = (
+            main.row_slice(0).unwrap(),
+            main.row_slice(1).unwrap(),
+        );
+        let (&alpha, beta_challenges) = challenges.split_first().expect("Wrong number of randomness");
+        let beta_challenges: [_; MAX_BETA_CHALLENGE_POWER] = beta_challenges.try_into().expect("Wrong number of randomness");
+        let periodic_values: [_; NUM_PERIODIC_VALUES] = periodic_evals.try_into().expect("Wrong number of periodic values");
+        vec![
+            (((alpha + beta_challenges[0]) * EF::from(main_current[0].clone()) + EF::ONE - EF::from(main_current[0].clone())) * EF::from(aux_current[0].clone())) * ((alpha + beta_challenges[0]) * (EF::from(main_current[0].clone()) - EF::ONE) + EF::ONE - (EF::from(main_current[0].clone()) - EF::ONE)).inverse(),
+            ((alpha + beta_challenges[0] + beta_challenges[1].double()) * (alpha + beta_challenges[0] + beta_challenges[1].double()) * (alpha + beta_challenges[0] + beta_challenges[1].double()) * EF::from(aux_current[1].clone()) + (alpha + beta_challenges[0] + beta_challenges[1].double()) * (alpha + beta_challenges[0] + beta_challenges[1].double()) * EF::from(main_current[0].clone()) + (alpha + beta_challenges[0] + beta_challenges[1].double()) * (alpha + beta_challenges[0] + beta_challenges[1].double()) * EF::from(main_current[0].clone()) - (alpha + beta_challenges[0] + beta_challenges[1].double()) * (alpha + beta_challenges[0] + beta_challenges[1].double()).double()) * ((alpha + beta_challenges[0] + beta_challenges[1].double()) * (alpha + beta_challenges[0] + beta_challenges[1].double()) * (alpha + beta_challenges[0] + beta_challenges[1].double())).inverse(),
+        ]
     }
 }