+++ air-script/src/tests/periodic_columns/periodic_columns_plonky3.rs
@@ -1,7 +1,8 @@
-use p3_air::{Air, BaseAir, BaseAirWithPublicValues, AirBuilder, ExtensionBuilder};
+use p3_field::{ExtensionField, Field, PrimeCharacteristicRing};
 use p3_matrix::Matrix;
-use p3_field::{Field, PrimeCharacteristicRing};
-use crate::test_utils::plonky3_traits::{AirScriptAir, AirScriptBuilder};
+use p3_matrix::dense::RowMajorMatrixView;
+use p3_matrix::stack::VerticalPair;
+use p3_miden_air::{MidenAir, MidenAirBuilder, RowMajorMatrix};
 
 pub const MAIN_WIDTH: usize = 3;
 pub const AUX_WIDTH: usize = 0;
@@ -12,25 +13,16 @@ pub const MAX_BETA_CHALLENGE_POWER: usize = 0;
 
 pub struct PeriodicColumnsAir;
 
-impl<F> BaseAir<F> for PeriodicColumnsAir {
+impl<F, EF> MidenAir<F, EF> for PeriodicColumnsAir
+where F: Field,
+      EF: ExtensionField<F>,
+{
     fn width(&self) -> usize {
         MAIN_WIDTH
     }
-}
 
-impl<F> BaseAirWithPublicValues<F> for PeriodicColumnsAir {
     fn num_public_values(&self) -> usize {
         NUM_PUBLIC_VALUES
-    }
-}
-
-impl<F: Field, AB: AirScriptBuilder<F = F>> AirScriptAir<F, AB> for PeriodicColumnsAir {
-    fn aux_width(&self) -> usize {
-        AUX_WIDTH
-    }
-
-    fn max_beta_challenge_power(&self) -> usize {
-        MAX_BETA_CHALLENGE_POWER
     }
 
     fn periodic_table(&self) -> Vec<Vec<F>> {
@@ -40,9 +32,12 @@ impl<F: Field, AB: AirScriptBuilder<F = F>> AirScriptAir<F, AB> for PeriodicColu
         ]
     }
 
-    fn eval(&self, builder: &mut AB) {
+    fn eval<AB>(&self, builder: &mut AB)
+    where AB: MidenAirBuilder<F = F, EF = EF>,
+    {
         let public_values: [_; NUM_PUBLIC_VALUES] = builder.public_values().try_into().expect("Wrong number of public values");
         let periodic_values: [_; NUM_PERIODIC_VALUES] = builder.periodic_evals().try_into().expect("Wrong number of periodic values");
+        let preprocessed = builder.preprocessed();
         let main = builder.main();
         let (main_current, main_next) = (
             main.row_slice(0).unwrap(),
@@ -53,17 +48,11 @@ impl<F: Field, AB: AirScriptBuilder<F = F>> AirScriptAir<F, AB> for PeriodicColu
         builder.when_first_row().assert_zero(main_current[0].clone().into());
 
         // Main integrity/transition constraints
-        builder.assert_zero_ext(periodic_values[0].into() * (AB::ExprEF::from(main_current[1].clone().into()) + AB::ExprEF::from(main_current[2].clone().into())));
-        builder.when_transition().assert_zero_ext(periodic_values[1].into() * (AB::ExprEF::from(main_next[0].clone().into()) - AB::ExprEF::from(main_current[0].clone().into())));
+        builder.assert_zero_ext(AB::ExprEF::from(periodic_values[0].clone().into()) * (AB::ExprEF::from(main_current[1].clone().into()) + AB::ExprEF::from(main_current[2].clone().into())));
+        builder.when_transition().assert_zero_ext(AB::ExprEF::from(periodic_values[1].clone().into()) * (AB::ExprEF::from(main_next[0].clone().into()) - AB::ExprEF::from(main_current[0].clone().into())));
 
         // Aux boundary constraints
 
         // Aux integrity/transition constraints
-    }
-}
-
-impl<AB: AirScriptBuilder> Air<AB> for PeriodicColumnsAir {
-    fn eval(&self, builder: &mut AB) {
-        <Self as AirScriptAir<AB::F, AB>>::eval(self, builder);
     }
 }