+++ codegen/plonky3/src/imports.rs
@@ -3,14 +3,13 @@ use super::Scope;
 /// Adds the required imports to the provided scope.
 pub(super) fn add_imports(scope: &mut Scope) {
     // add plonky3 imports
-    scope.import("p3_air", "Air");
-    scope.import("p3_air", "BaseAir");
-    scope.import("p3_air", "BaseAirWithPublicValues");
-    scope.import("p3_air", "AirBuilder");
-    scope.import("p3_air", "ExtensionBuilder");
-    scope.import("p3_matrix", "Matrix");
+    scope.import("p3_field", "ExtensionField");
     scope.import("p3_field", "Field");
     scope.import("p3_field", "PrimeCharacteristicRing");
-    scope.import("crate::test_utils::plonky3_traits", "AirScriptAir");
-    scope.import("crate::test_utils::plonky3_traits", "AirScriptBuilder");
+    scope.import("p3_matrix", "Matrix");
+    scope.import("p3_matrix::dense", "RowMajorMatrixView");
+    scope.import("p3_matrix::stack", "VerticalPair");
+    scope.import("p3_miden_air", "MidenAir");
+    scope.import("p3_miden_air", "MidenAirBuilder");
+    scope.import("p3_miden_air", "RowMajorMatrix");
 }